ReadReccomend/src/main/requirements.txt
aniso8601==8.0.0
asn1crypto==0.24.0
attrs==19.3.0
Automat==0.6.0
backcall==0.2.0
bleach==3.1.5
blinker==1.4
certifi==2018.1.18
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
cryptography==2.9.2
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.2
flask-restplus==0.13.0
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
importlib-metadata==1.6.1
incremental==16.10.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
joblib==0.15.1
jsonpatch==1.16
jsonpointer==1.10
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
keyring==10.6.0
keyrings.alt==3.0
liac-arff==2.4.0
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
netifaces==0.10.4
notebook==6.0.3
numpy==1.19.0
oauthlib==2.0.6
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
pycrypto==2.6.1
Pygments==2.6.1
PyJWT==1.7.1
pyOpenSSL==17.5.0
pyparsing==2.4.7
pyrsistent==0.16.0
pyserial==3.4
python-dateutil==2.8.1
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
requests==2.18.4
requests-unixsocket==0.1.5
scikit-learn==0.23.1
scipy==1.5.0
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==16.0.0
simplegeneric==0.8.1
six==1.15.0
sklearn==0.0
ssh-import-id==5.7
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
traitlets==4.3.3
Twisted==17.9.0
urllib3==1.22
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
zipp==3.1.0